Move component types to their own file
