+++ src/AlignHorizontal/index.ts
@@ -1,1 +1,2 @@
 export { default } from "./AlignHorizontal";
+export { AlignHorizontalProps } from "./AlignHorizontal.types";
